requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
